Move movement of code out of Eagle.ts
