+++ static/components/repository-file.html
@@ -16,7 +16,7 @@
         </div>
         <!-- ko if: getIconUrl() === "device_hub" -->
             <div class="col col-1">
-                <i class="material-icons md-18" data-bs-placement="left" style="cursor:pointer;top:3px;" data-bind="click: $root.insertFile, eagleTooltip: 'Insert'">add_box</i>
+                <i class="material-icons md-18" data-bs-placement="left" style="cursor:pointer;top:3px;" data-bind="click: $root.insertRemoteFile, eagleTooltip: 'Insert'">add_box</i>
             </div>
         <!-- /ko -->
     </div>
