+++ templates/repositories.html
@@ -10,7 +10,7 @@
     </div>
     <div class="row">
         <div class="col">
-            <button type="button" id="addRepository" class="btn btn-primary btn-block btn_wide" data-bind="click: addCustomRepository">Add Repository</button>
+            <button type="button" id="addRepository" class="btn btn-primary btn-block btn_wide" data-bind="click: Repositories.addCustomRepository">Add Repository</button>
         </div>
     </div>
 </div>
