+++ tsconfig.json
@@ -47,6 +47,7 @@
         "./src/ExplorePalettes.ts",
         "./src/Undo.ts",
         "./src/Errors.ts",
-        "./src/Hierarchy.ts"
+        "./src/Hierarchy.ts",
+        "./src/Repositories.ts"
     ]
 }
